backend/requirements.txt
uvicorn==0.30.6
fastapi==0.112.2
pydantic==2.10.4
SQLAlchemy==2.0.36
python-dotenv==1.0.1
